+++ packages/cli/zp.config.js
@@ -93,11 +93,12 @@ const config = {
   backtestsDir: "./dev/backtests",
   dataProvider: "zapant", // async (params) => [{...}]
   execute: {
-    date: "2024-04-01",
+    date: "2024-04-07",
     errors: "./dev/errors.txt",
     inputs: {
       assets: [],
       initialCapital: 10000,
+      cash: 5000,
       openPositions: [] // samplePositions
     },
   },