+++ packages/core/src/trading/core.ts
@@ -1,4 +1,4 @@
-import { pluck, head, take } from 'ramda'
+import { pluck, head, take, tail } from 'ramda'
 
 const zpCore = (env) => {
   const { stdout } = env
@@ -7,6 +7,7 @@ const zpCore = (env) => {
     pluck,
     head,
     take,
+    tail,
     print: (...args) => {
       const str = args.map(x => typeof x === 'string' ? x : JSON.stringify(x, null, 2)).join(' ')
       stdout.push(str)